+++ x-pack/solutions/observability/plugins/apm/public/components/app/storage_explorer/get_storage_explorer_links.ts
@@ -14,10 +14,6 @@ export function getIndexManagementHref(core: CoreStart, dataStream?: string) {
   });
 }
 
-export function getStorageExplorerFeedbackHref() {
-  return 'https://ela.st/feedback-storage-explorer';
-}
-
 export function getKibanaAdvancedSettingsHref(core: CoreStart) {
   return core.application.getUrlForApp('management', {
     path: '/kibana/settings?query=categories:(observability)',
