+++ x-pack/solutions/observability/plugins/apm/public/components/app/storage_explorer/resources/tips_and_resources.tsx
@@ -21,16 +21,12 @@ import { i18n } from '@kbn/i18n';
 import { useApmRouter } from '../../../../hooks/use_apm_router';
 import { useApmParams } from '../../../../hooks/use_apm_params';
 import { useApmPluginContext } from '../../../../context/apm_plugin/use_apm_plugin_context';
-import {
-  getIndexManagementHref,
-  getStorageExplorerFeedbackHref,
-} from '../get_storage_explorer_links';
+import { getIndexManagementHref } from '../get_storage_explorer_links';
 
 export function TipsAndResources() {
   const router = useApmRouter();
   const { core } = useApmPluginContext();
   const { docLinks } = core;
-  const isFeedbackEnabled = core.notifications.feedback.isEnabled();
 
   const {
     query: { rangeFrom, rangeTo, environment, kuery, comparisonEnabled },
@@ -107,18 +103,6 @@ export function TipsAndResources() {
       target: '_blank',
       iconType: 'documentation',
     },
-    ...(isFeedbackEnabled
-      ? [
-          {
-            label: i18n.translate('xpack.apm.storageExplorer.resources.sendFeedback', {
-              defaultMessage: 'Give feedback',
-            }),
-            href: getStorageExplorerFeedbackHref(),
-            target: '_blank',
-            iconType: 'comment',
-          },
-        ]
-      : []),
   ];
 
   return (
