+++ x-pack/solutions/observability/plugins/infra/public/components/asset_details/tabs/anomalies/anomalies.tsx
@@ -5,14 +5,11 @@
  * 2.0.
  */
 
-import React, { useContext, useMemo, useRef } from 'react';
-import { FeatureFeedbackButton } from '@kbn/observability-shared-plugin/public';
+import React, { useMemo, useRef } from 'react';
 import { AnomaliesTable } from '../../../ml/anomaly_detection/anomalies_table/anomalies_table';
 import { useAssetDetailsRenderPropsContext } from '../../hooks/use_asset_details_render_props';
 import { useDatePickerContext } from '../../hooks/use_date_picker';
 import { useRequestObservable } from '../../hooks/use_request_observable';
-import { KibanaEnvironmentContext } from '../../../../hooks/use_kibana';
-import { INFRA_ML_FLYOUT_FEEDBACK_LINK } from '../../../ml/anomaly_detection/flyout_home';
 import { useTabSwitcherContext } from '../../hooks/use_tab_switcher';
 
 export const Anomalies = () => {
@@ -21,8 +18,6 @@ export const Anomalies = () => {
   const { request$ } = useRequestObservable();
   const { getParsedDateRange } = useDatePickerContext();
   const { entity } = useAssetDetailsRenderPropsContext();
-  const { kibanaVersion, isCloudEnv, isServerlessEnv } = useContext(KibanaEnvironmentContext);
-
   const parsedDateRange = useMemo(() => getParsedDateRange(), [getParsedDateRange]);
 
   return (
@@ -36,26 +31,6 @@ export const Anomalies = () => {
           requestObservable$: request$,
         }}
       />
-      <div
-        style={{
-          position: 'fixed',
-          bottom: '24px',
-          right: '24px',
-          display: 'flex',
-          justifyContent: 'flex-end',
-          alignItems: 'flex-end',
-        }}
-      >
-        <FeatureFeedbackButton
-          data-test-subj="infraMLHostFlyoutFeedbackLink"
-          formUrl={INFRA_ML_FLYOUT_FEEDBACK_LINK}
-          sanitizedPath={document.location.pathname}
-          kibanaVersion={kibanaVersion}
-          isCloudEnv={isCloudEnv}
-          isServerlessEnv={isServerlessEnv}
-          nodeType="host"
-        />
-      </div>
     </div>
   );
 };
