+++ x-pack/solutions/observability/plugins/observability_shared/public/index.ts
@@ -95,10 +95,6 @@ export {
 } from './components/profiling/embeddables';
 
 export { ProfilingEmptyState } from './components/profiling/profiling_empty_state';
-export {
-  FeatureFeedbackButton,
-  getSurveyFeedbackURL,
-} from './components/feature_feedback_button/feature_feedback_button';
 export { BottomBarActions } from './components/bottom_bar_actions/bottom_bar_actions';
 export { FieldValueSelection, FieldValueSuggestions } from './components';
