+++ x-pack/solutions/observability/plugins/profiling/public/components/frame_information_window/missing_symbols_callout.tsx
@@ -11,7 +11,6 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import { i18n } from '@kbn/i18n';
 import type { FrameType } from '@kbn/profiling-utils';
 import { getLanguageType } from '@kbn/profiling-utils';
-import { PROFILING_FEEDBACK_LINK } from '../profiling_app_page_template';
 import { useProfilingDependencies } from '../contexts/profiling_dependencies/use_profiling_dependencies';
 import { useProfilingRouter } from '../../hooks/use_profiling_router';
 import { AddDataTabs } from '../../views/add_data_view';
@@ -23,8 +22,7 @@ interface Props {
 export function MissingSymbolsCallout({ frameType }: Props) {
   const languageType = getLanguageType({ frameType });
   const router = useProfilingRouter();
-  const { docLinks, notifications } = useProfilingDependencies().start.core;
-  const isFeedbackEnabled = notifications.feedback.isEnabled();
+  const { docLinks } = useProfilingDependencies().start.core;
 
   if (languageType === 'NATIVE') {
     return (
@@ -88,19 +86,6 @@ export function MissingSymbolsCallout({ frameType }: Props) {
             'Symbols are not available because of an error in the unwinder for this language or an unknown error with the interpreter.',
         })}
       </p>
-      {isFeedbackEnabled && (
-        <EuiButton
-          data-test-subj="profilingMissingSymbolsCalloutReportAProblemButton"
-          href={PROFILING_FEEDBACK_LINK}
-          target="_blank"
-          color="warning"
-        >
-          {i18n.translate(
-            'xpack.profiling.frameInformationWindow.missingSymbols.interpreted.reportProblem',
-            { defaultMessage: 'Report a problem' }
-          )}
-        </EuiButton>
-      )}
     </EuiCallOut>
   );
 }
