+++ x-pack/solutions/observability/plugins/profiling/public/components/profiling_app_page_template/index.tsx
@@ -23,8 +23,6 @@ import { PrimaryProfilingSearchBar } from './primary_profiling_search_bar';
 import { useLocalStorage } from '../../hooks/use_local_storage';
 import { useProfilingSetupStatus } from '../contexts/profiling_setup_status/use_profiling_setup_status';
 
-export const PROFILING_FEEDBACK_LINK = 'https://ela.st/profiling-feedback';
-
 export function ProfilingAppPageTemplate({
   children,
   tabs = [],
@@ -45,7 +43,7 @@ export function ProfilingAppPageTemplate({
   showBetaBadge?: boolean;
 }) {
   const {
-    start: { observabilityShared, core },
+    start: { observabilityShared },
   } = useProfilingDependencies();
 
   const [privilegesWarningDismissed, setPrivilegesWarningDismissed] = useLocalStorage(
@@ -58,8 +56,6 @@ export function ProfilingAppPageTemplate({
 
   const history = useHistory();
 
-  const isFeedbackEnabled = core.notifications.feedback.isEnabled();
-
   useEffect(() => {
     window.scrollTo(0, 0);
   }, [history.location.pathname]);
@@ -69,21 +65,6 @@ export function ProfilingAppPageTemplate({
       noDataConfig={noDataConfig}
       pageHeader={{
         'data-test-subj': 'profilingPageTemplate',
-        rightSideItems: isFeedbackEnabled
-          ? [
-              <EuiButton
-                data-test-subj="profilingProfilingAppPageTemplateGiveFeedbackButton"
-                href={PROFILING_FEEDBACK_LINK}
-                target="_blank"
-                color="warning"
-                iconType="comment"
-              >
-                {i18n.translate('xpack.profiling.header.giveFeedbackLink', {
-                  defaultMessage: 'Give feedback',
-                })}
-              </EuiButton>,
-            ]
-          : undefined,
         pageTitle: (
           <EuiFlexGroup gutterSize="s" alignItems="baseline">
             <EuiFlexItem grow={false}>{pageTitle}</EuiFlexItem>
